# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ServiceFlow SDM API
  version: 1.0.0
  description: |
    Service Delivery Manager backend API for MGM service delivery operations.

    ## Authentication
    All protected endpoints require a Bearer JWT access token obtained from
    `POST /api/v1/auth/login` or `POST /api/v1/auth/sso/exchange`.

    ## Roles
    Roles in order of privilege: `admin > sdm > manager > tam > csm > engineer > viewer`

    `viewer` has read-only access scoped to their assigned properties/incidents.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Unauthorized
...
    TokenPair:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    Device:
      type: object
      required: [device_id, property_id, hostname, ip_address, serial_number, status]
      properties:
        device_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        hostname:
          type: string
          example: sw-bellagio-core-01
        ip_address:
          type: string
          format: ipv4
          example: 10.1.1.1
        serial_number:
          type: string
          example: FCZ2341A0AB
        status:
          type: string
          enum: [active, inactive, maintenance, decommissioned, failed]
        dna_device_id:
          type: string
          nullable: true
        dna_managed:
          type: boolean
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      summary: Login with email and password
      description: Returns an access token (15 min) and a refresh token (7 days).
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded (20 attempts per 15 minutes)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Exchanges a valid refresh token for a new access token.
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                required: [accessToken]
                properties:
                  accessToken:
                    type: string
        "400":
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/sso/exchange:
    post:
      summary: Exchange SSO one-time code for tokens
      description: |
        Called by the frontend immediately after the SSO redirect. The `code` is
        single-use (Redis GETDEL) and expires after 60 seconds.
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Missing code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Code invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Properties ─────────────────────────────────────────────────────────────
...
  /devices:
    get:
      summary: List devices
      description: |
        Privileged roles see all devices. `viewer` role sees only devices belonging
        to their assigned properties (`mgm.user_properties`).
      tags: [Devices]
      responses:
        "200":
          description: Array of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a device
      description: Requires role `admin`, `sdm`, `engineer`, or `manager`.
      tags: [Devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propertyId, hostname, ipAddress, serialNumber, status]
              properties:
                propertyId:
                  type: string
                  format: uuid
                hostname:
                  type: string
                ipAddress:
                  type: string
                  format: ipv4
                serialNumber:
                  type: string
                  minLength: 3
                status:
                  type: string
                  enum: [active, inactive, maintenance, decommissioned, failed]
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /devices/{deviceId}:
    get:
      summary: Get a single device
      description: Viewers receive 404 if the device's property is not in their assignment.
      tags: [Devices]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Device object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Incidents ──────────────────────────────────────────────────────────────
...